# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Artifact Repository API
  version: 1.0.0
  description: API-first immutable binary artifact repository.
servers:
  - url: /
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ar1
  parameters:
    Repo:
      x-go-name: RepoPathParameter
      name: repo
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Name"
    PackageName:
      x-go-name: PackagePathParameter
      name: package
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Name"
...
    Version:
      x-go-name: VersionPathParameter
      name: version
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Version"
    ArtifactPath:
      x-go-name: ArtifactPathParameter
      name: path
      in: path
      required: true
      allowReserved: true
      description: Normalized slash-separated logical path; runtime routing captures remaining segments.
      schema:
        type: string
        minLength: 1
        maxLength: 1024
    ReleaseArtifactID:
      x-go-name: ReleaseArtifactIDPathParameter
      name: releaseArtifactId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      x-go-name: CursorQueryParameter
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 512
    Limit:
      x-go-name: LimitQueryParameter
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    IdempotencyKey:
      x-go-name: IdempotencyKeyHeaderParameter
      name: Idempotency-Key
      in: header
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]+$"
    ContentLength:
      x-go-name: ContentLengthHeaderParameter
      name: Content-Length
      in: header
      required: true
      schema:
        type: integer
        format: int64
        minimum: 0
        maximum: 10737418240
    ChecksumSha256:
      x-go-name: ChecksumSha256HeaderParameter
      name: X-Checksum-Sha256
      in: header
      schema:
        $ref: "#/components/schemas/Sha256"
    ChecksumDeploy:
      x-go-name: ChecksumDeployHeaderParameter
      name: X-Checksum-Deploy
      in: header
      schema:
        type: boolean
        default: false
    ArtifactProperties:
      x-go-name: ArtifactPropertiesHeaderParameter
      name: X-Artifact-Properties
      in: header
      description: Base64URL-encoded UTF-8 JSON Object, decoded maximum 16 KiB.
      schema:
        type: string
        maxLength: 21846
        pattern: "^[A-Za-z0-9_-]*$"
    Redirect:
      x-go-name: RedirectQueryParameter
      name: redirect
      in: query
      schema:
        type: boolean
        default: true
    Range:
      x-go-name: RangeHeaderParameter
      name: Range
      in: header
      schema:
        type: string
        maxLength: 256
        pattern: "^bytes="
  responses:
    BadRequest:
      description: Request syntax or validation failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Authentication failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Scope or Repository authorization failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource was not found or is intentionally hidden.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Immutable path, state, uniqueness, or idempotency conflict.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unprocessable:
      description: Domain ownership or checksum constraint failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PayloadTooLarge:
      description: Upload exceeds the configured maximum.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    TooManyRequests:
      description: The per-Token endpoint-class limit or upload concurrency limit was exceeded.
      headers:
        Retry-After:
          required: true
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InternalError:
      description: Non-retryable internal failure.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ServiceUnavailable:
      description: Dependency unavailable or operation retained for recovery.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Name:
      type: string
      minLength: 2
      maxLength: 64
      pattern: "^[a-z][a-z0-9._-]{1,63}$"
...
    Version:
      type: string
      minLength: 1
      maxLength: 128
      pattern: '^(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)(-((0|[1-9][0-9]*|[0-9]*[A-Za-z-][0-9A-Za-z-]*)(\.(0|[1-9][0-9]*|[0-9]*[A-Za-z-][0-9A-Za-z-]*))*))?(\+([0-9A-Za-z-]+(\.[0-9A-Za-z-]+)*))?$'
    Coordinate:
      type: string
      minLength: 1
      maxLength: 64
      pattern: "^[A-Za-z0-9][A-Za-z0-9._+-]{0,63}$"
    OptionalCoordinate:
      type: string
      maxLength: 64
      pattern: "^[A-Za-z0-9._+-]{0,64}$"
      default: ""
    Sha256:
      type: string
      pattern: "^[0-9a-f]{64}$"
    UUID:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          maxLength: 256
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          minLength: 1
          maxLength: 128
        detail:
          type: string
          maxLength: 2048
        requestId:
          type: string
          minLength: 1
          maxLength: 64
    PageCursor:
      type: string
      maxLength: 512
      nullable: true
...
    Artifact:
      type: object
      additionalProperties: false
      required: [id, repository, path, filename, mediaType, size, sha256, properties, createdBy, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        repository:
          $ref: "#/components/schemas/Name"
        path:
          type: string
          minLength: 1
          maxLength: 1024
        filename:
          type: string
          minLength: 1
          maxLength: 255
        mediaType:
          type: string
          maxLength: 255
        size:
          type: integer
          format: int64
          minimum: 0
        sha256:
          $ref: "#/components/schemas/Sha256"
        properties:
          type: object
          maxProperties: 256
          additionalProperties: true
        createdBy:
          $ref: "#/components/schemas/UUID"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
...
    CreateReleaseRequest:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version:
          $ref: "#/components/schemas/Version"
    ReleaseState:
      type: string
      enum: [draft, publishing, published, publish_failed]
    InstallHook:
      type: object
      additionalProperties: false
      required: [phase, path, timeoutSeconds]
      properties:
        phase:
          type: string
          enum: [preflight, post-install, verify]
        path:
          type: string
          minLength: 1
          maxLength: 1024
        args:
          type: array
          maxItems: 16
          items:
            type: string
            minLength: 1
            maxLength: 1024
        timeoutSeconds:
          type: integer
          minimum: 1
          maximum: 300
    InstallSpec:
      type: object
      additionalProperties: false
      required: [strategy, format, mode]
      properties:
        strategy:
          type: string
          enum: [self-replace, bundle]
        format:
          type: string
          enum: [raw, tar.gz, zip]
        entrypoint:
          type: string
          maxLength: 1024
        mode:
          type: string
          pattern: "^0[0-7]{3}$"
        hooks:
          type: array
          maxItems: 3
          items:
            $ref: "#/components/schemas/InstallHook"
    AddReleaseArtifactRequest:
      type: object
      additionalProperties: false
      required: [artifactPath, os, arch]
      properties:
        artifactPath:
          type: string
          minLength: 1
          maxLength: 1024
        os:
          $ref: "#/components/schemas/Coordinate"
        arch:
          $ref: "#/components/schemas/Coordinate"
        variant:
          $ref: "#/components/schemas/OptionalCoordinate"
        role:
          $ref: "#/components/schemas/OptionalCoordinate"
          description: Product/installable releases must use exactly "binary"; the field remains optional for legacy non-product releases.
        install:
          $ref: "#/components/schemas/InstallSpec"
    ReleaseArtifact:
      type: object
      additionalProperties: false
      required: [id, artifact, os, arch, variant, role]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        artifact:
          $ref: "#/components/schemas/Artifact"
        os:
          $ref: "#/components/schemas/Coordinate"
        arch:
          $ref: "#/components/schemas/Coordinate"
        variant:
          $ref: "#/components/schemas/OptionalCoordinate"
        role:
          $ref: "#/components/schemas/OptionalCoordinate"
        install:
          $ref: "#/components/schemas/InstallSpec"
    Release:
      type: object
      additionalProperties: false
      required: [id, repository, package, version, state, artifacts, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        repository:
          $ref: "#/components/schemas/Name"
        package:
          $ref: "#/components/schemas/Name"
        version:
          $ref: "#/components/schemas/Version"
        state:
          $ref: "#/components/schemas/ReleaseState"
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/ReleaseArtifact"
        publishedAt:
          $ref: "#/components/schemas/Timestamp"
        failureCode:
          type: string
          maxLength: 128
        createdAt:
          $ref: "#/components/schemas/Timestamp"
    ReleasePage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Release"
        nextCursor:
          $ref: "#/components/schemas/PageCursor"
    ReleaseManifest:
      type: object
      additionalProperties: false
      required: [version, manifest, manifestSha256, keyId, signature]
      properties:
        version:
          $ref: "#/components/schemas/Version"
        manifest:
          type: string
          contentEncoding: base64url
        manifestSha256:
          $ref: "#/components/schemas/Sha256"
        keyId:
          type: string
          minLength: 8
          maxLength: 128
        signature:
          type: string
          contentEncoding: base64url
...
paths:
...
  /api/v1/repositories/{repo}/artifacts/{path}:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/ArtifactPath"
    put:
      operationId: putArtifact
      parameters:
        - $ref: "#/components/parameters/ContentLength"
        - $ref: "#/components/parameters/ChecksumSha256"
        - $ref: "#/components/parameters/ChecksumDeploy"
        - $ref: "#/components/parameters/ArtifactProperties"
      requestBody:
        required: false
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Immutable Artifact created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "429":
          $ref: '#/components/responses/TooManyRequests'
    head:
      operationId: headArtifact
      responses:
        "200":
          description: Artifact headers.
          headers:
            Content-Length:
              schema:
                type: integer
                format: int64
            ETag:
              schema:
                type: string
            X-Checksum-Sha256:
              schema:
                $ref: "#/components/schemas/Sha256"
            X-Created-At:
              schema:
                type: string
                format: date-time
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: '#/components/responses/TooManyRequests'
    get:
      operationId: getArtifact
      parameters:
        - $ref: "#/components/parameters/Redirect"
        - $ref: "#/components/parameters/Range"
      responses:
        "200":
          description: Proxied artifact bytes.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "206":
          description: Proxied byte range.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "307":
          description: Redirect to a short-lived public object-storage URL when supported by the selected backend.
          headers:
            Location:
              required: true
              schema:
                type: string
                format: uri
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: '#/components/responses/TooManyRequests'
...
  /api/v1/repositories/{repo}/packages/{package}/releases:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/PackageName"
    post:
      operationId: createRelease
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReleaseRequest"
      responses:
        "201":
          description: Draft Release created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: '#/components/responses/TooManyRequests'
    get:
      operationId: listReleases
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Releases in the Package.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleasePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/repositories/{repo}/packages/{package}/releases/{version}:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/PackageName"
      - $ref: "#/components/parameters/Version"
    get:
      operationId: getRelease
      responses:
        "200":
          description: Release details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: '#/components/responses/TooManyRequests'
    delete:
      operationId: cancelRelease
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204":
          description: Draft Release canceled.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/repositories/{repo}/packages/{package}/releases/{version}/artifacts:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/PackageName"
      - $ref: "#/components/parameters/Version"
    post:
      operationId: addReleaseArtifact
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddReleaseArtifactRequest"
      responses:
        "201":
          description: Artifact reference added.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseArtifact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "429":
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/repositories/{repo}/packages/{package}/releases/{version}/artifacts/{releaseArtifactId}:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/PackageName"
      - $ref: "#/components/parameters/Version"
      - $ref: "#/components/parameters/ReleaseArtifactID"
    delete:
      operationId: removeReleaseArtifact
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204":
          description: Draft Artifact reference removed.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/repositories/{repo}/packages/{package}/releases/{version}/publish:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/PackageName"
      - $ref: "#/components/parameters/Version"
    post:
      operationId: publishRelease
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Release published.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "429":
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/repositories/{repo}/packages/{package}/releases/{version}/manifest:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/PackageName"
      - $ref: "#/components/parameters/Version"
    get:
      operationId: getReleaseManifest
      responses:
        "200":
          description: Exact signed Manifest.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseManifest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: '#/components/responses/TooManyRequests'